requirements.txt
django-image-cropping==1.0.3
easy_thumbnails==2.3
